dm vdo test: allow thread count changes
Add functionality to adjust LVM's VDO properties when the
VDO volume is offline (between restarts).

Tests used to be able to change thread counts and other
configuration properties when restarting a VDO device.
When we switched to using LVM, we commented those calls
out because LVM, at the time, did not have a way to
change VDO parameters.

However, LVM added that functionality a few years back,
so now we can add a command to do the necessary lvchange
operation, and call it in tests that used to that. The
new pendingSettings hash can store updated values that
will be applied to the VDO volume the next time it is
restarted.

Signed-off-by: Matthew Sakai <msakai@redhat.com>